Fix integer overflow in OptionConverter (enable 64-bit sizes on Windows)
Changed toFileSize signature to return 'long long' instead of 'long'.
On Windows (LLP64), 'long' is 32-bit, causing overflow for sizes >= 2GB.
Using 'long long' guarantees 64-bit width, preventing negative file sizes
and enabling support for large log files.

+++ src/main/include/log4cxx/helpers/optionconverter.h
@@ -72,7 +72,8 @@ class LOG4CXX_EXPORT OptionConverter
 		 converts the provided number respectively to kilobytes, megabytes
 		 and gigabytes. For example, the value "10KB" will be interpreted as 10240.
 		*/
-		static long toFileSize(const LogString& value, long defaultValue);
+		// Use long long to ensure 64-bit capacity on Windows (LLP64) and Linux (LP64)
+                static long long toFileSize(const LogString& value, long long defaultValue);
 		/**
 		The Level indicated by \c value if recognised otherwise \c defaultValue.
